refactor: 도메인 로직은 domain/ 하위, UI 관련 로직은 view/ 하위에서 관리하도록 수정
