+++ docs/REQUIREMENTS.md
@@ -64,4 +64,8 @@
 
 - [x] randomGenerator 를 도메인과 분리시켜 보다 더 유틸화하기
 - [x] for문을 Array의 메소드로 대체하기
+- [x] 도메인 로직은 domain/ 하위로, UI 관련 로직은 view/ 하위에서 관리하기
+- [x] domain/ 하위의 모듈은 view/ 하위의 모듈을 의존하지 않아야 한다 (컨트롤러에
+      서 입출력 명령을 담당하기 때문에 도메인 로직이 UI 관련 로직을 알일은 없지
+      않나 (?)
 - [] 정상적인 입력 값이 들어왔을 경우의 테스트 추가
